% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{cox.prediction}
\alias{cox.prediction}
\title{cox.prediction}
\usage{
cox.prediction(model, new_data, time = NULL, type = "lp", method = "cox")
}
\arguments{
\item{model}{HDcox model.}

\item{new_data}{Numeric matrix or data.frame. New explanatory variables (raw data). Qualitative variables must be transform into binary variables.}

\item{time}{Numeric. Time point where the AUC will be evaluated.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{method}{Character. Prediction method. It can be compute by using the cox model "cox" or by using W.star "W.star" (default: "cox"). (not implemented for MB approaches)!!!}
}
\value{
Return the lp or other metric for the patient
}
\description{
cox.prediction
}
