% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{cox_comparePatients}
\alias{cox_comparePatients}
\title{cox.comparePatients}
\usage{
cox_comparePatients(
  model,
  df.pat,
  error.bar = F,
  onlySig = T,
  alpha = 0.05,
  zero.rm = T,
  auto.limits = T,
  top = NULL
)
}
\arguments{
\item{model}{HDcox model}

\item{df.pat}{Dataframe of observations}

\item{error.bar}{Show error bar.}

\item{onlySig}{Show only significant variables.}

\item{alpha}{Significant value.}

\item{zero.rm}{Remove variables equal to 0.}

\item{auto.limits}{Logical. If TRUE, limits are detected for a better plot.}

\item{top}{Number. Show top variables.}
}
\description{
cox.comparePatients
}
