% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_coxSW.R
\name{coxSW}
\alias{coxSW}
\title{coxSW}
\usage{
coxSW(
  X,
  Y,
  x.center = TRUE,
  x.scale = FALSE,
  y.center = FALSE,
  y.scale = FALSE,
  remove_near_zero_variance = T,
  remove_zero_variance = F,
  toKeep.zv = NULL,
  remove_non_significant = F,
  initialModel = "NULL",
  toKeep.sw = NULL,
  max.variables = 20,
  alpha = 0.05,
  alpha_ENT = 0.1,
  alpha_OUT = 0.15,
  alpha_PH = 0.05,
  check_PH = F,
  boostDeletion = F,
  BACKWARDS = T,
  MIN_EPV = 5,
  returnData = T,
  verbose = F
)
}
\arguments{
\item{X}{Numeric matrix or data.frame. Explanatory variables. Qualitative variables must be transform into binary variables.}

\item{Y}{Numeric matrix or data.frame. Response variables. Object must have two columns named as "time" and "event". For event column, accepted values are: 0/1 or FALSE/TRUE for censored and event observations.}

\item{x.center}{Logical. If x.center = TRUE, X matrix is centered to zero means (default: TRUE).}

\item{x.scale}{Logical. If x.scale = TRUE, X matrix is scaled to unit variances (default: FALSE).}

\item{y.center}{Logical. If y.center = TRUE, Y matrix is centered to zero means (default: FALSE).}

\item{y.scale}{Logical. If y.scale = TRUE, Y matrix is scaled to unit variances (default: FALSE).}

\item{remove_near_zero_variance}{Logical. If remove_near_zero_variance = TRUE, near zero variance variables will be removed (default: TRUE).}

\item{remove_zero_variance}{Logical. If remove_zero_variance = TRUE, zero variance variables will be removed (default: TRUE).}

\item{toKeep.zv}{Character vector. Name of variables in X to not be deleted by (near) zero variance filtering (default: NULL).}

\item{remove_non_significant}{Logical. If remove_non_significant = TRUE, non-significant variables/components in final cox model will be removed until all variables are significant by forward selection (default: FALSE).}

\item{initialModel}{Character vector. Name of variables in X to include in the initial model (default: NULL).}

\item{toKeep.sw}{Character vector. Name of variables in X to not be deleted by Step-wise selection (default: NULL).}

\item{max.variables}{Numeric. Maximum number of variables you want to keep in the cox model. If MIN_EPV is not meet, the value will be change automatically (default: 20).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{alpha_ENT}{Numeric. Maximum P-Value for a variable to enter the model (default: 0.10).}

\item{alpha_OUT}{Numeric. Minimum P-Value for a variable to leave the model (default: 0.15).}

\item{alpha_PH}{Numeric. Maximum P-Value for a variable to meet Proportional Hazard Assumption (default: 0.05).}

\item{check_PH}{Logical. If check_PH = TRUE, PH Assumption will be check (default: FALSE).}

\item{boostDeletion}{Logical. If boostDeletion = TRUE, all variables with a P-Value greater than alpha_OUT will be deleted in each step, instead of deleting one by one (default: FALSE).}

\item{BACKWARDS}{Logical. If BACKWARDS = TRUE, backward strategy is performed (default: TRUE).}

\item{MIN_EPV}{Numeric. Minimum number of Events Per Variable (EPV) you want reach for the final cox model. Used to restrict the number of variables/components can be computed in final cox models. If the minimum is not meet, the model cannot be computed (default: 5).}

\item{returnData}{Logical. Return original and normalized X and Y matrices (default: TRUE).}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}
}
\value{
Instance of class "HDcox" and model "coxSW". The class contains the following elements:

\code{X}: List of normalized X data information.
\itemize{
\item \code{(data)}: normalized X matrix
\item \code{(x.mean)}: mean values for X matrix
\item \code{(x.sd)}: standard deviation for X matrix
}
\code{Y}: List of normalized Y data information.
\itemize{
\item \code{(data)}: normalized Y matrix
\item \code{(y.mean)}: mean values for Y matrix
\item \code{(y.sd)}: standard deviation for Y matrix
}
\code{survival_model}: List of survival model information
\itemize{
\item \code{fit}: coxph object.
\item \code{AIC}: AIC of cox model.
\item \code{BIC}: BIC of cox model.
\item \code{lp}: linear predictors for train data.
\item \code{coef}: Coefficients for cox model.
\item \code{YChapeau}: Y Chapeau residuals.
\item \code{Yresidus}: Y residuals.
}
\code{call}: call function

\code{X_input}: X input matrix

\code{Y_input}: Y input matrix

\code{nzv}: Variables removed by remove_near_zero_variance or remove_zero_variance.

\code{class}: Model class.

\code{time}: time consumed for running the cox analysis.
}
\description{
This function performs a cox stepwise model (based on My.stepwise R package).
The function returns a HDcox model with the attribute model as "coxSW".
}
\examples{
\dontrun{
coxSW(X, Y)
coxSW(X, Y, x.center = TRUE, x.scale = TRUE)
}
}
