% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_splsdacox_mixomics.R
\name{cv.splsdacox_mixOmics}
\alias{cv.splsdacox_mixOmics}
\title{Cross validation splsdacox_mixOmics}
\usage{
cv.splsdacox_mixOmics(
  X,
  Y,
  max.ncomp = 10,
  n_run = 10,
  k_folds = 10,
  vector = NULL,
  x.center = TRUE,
  x.scale = FALSE,
  y.center = FALSE,
  y.scale = FALSE,
  remove_near_zero_variance = T,
  remove_zero_variance = T,
  toKeep.zv = NULL,
  remove_non_significant_models = F,
  alpha = 0.05,
  MIN_NVAR = 10,
  MAX_NVAR = 1000,
  n.cut_points = 5,
  MIN_AUC_INCREASE = 0.01,
  EVAL_METHOD = "AUC",
  w_AIC = 0,
  w_c.index = 0,
  w_AUC = 1,
  times = NULL,
  MIN_AUC = 0.8,
  MIN_COMP_TO_CHECK = 3,
  pred.attr = "mean",
  pred.method = "cenROC",
  fast_mode = F,
  max.iter = 500,
  MIN_EPV = 5,
  return_models = F,
  PARALLEL = F,
  verbose = F,
  seed = 123
)
}
\arguments{
\item{X}{Numeric matrix. Predictor variables}

\item{Y}{Numeric matrix. Response variables. It assumes it has two columns named as "time" and "event". For event column, values can be 0/1 or FALSE/TRUE for censored and event samples.}

\item{max.ncomp}{Numeric. Maximum number of PLS components to compute for the cross validation.}

\item{n_run}{Number. Number of runs for cross validation.}

\item{k_folds}{Number. Number of folds for cross validation.}

\item{vector}{Numeric vector. Used for computing best number of variables. If NULL, an automatic detection is perform.}

\item{x.center}{Logical. If x.center = TRUE, X matrix is centered to zero means (default: TRUE).}

\item{x.scale}{Logical. If x.scale = TRUE, X matrix is scaled to unit variances (default: FALSE).}

\item{y.center}{Logical. If y.center = TRUE, Y matrix is centered to zero means (default: FALSE).}

\item{y.scale}{Logical. If y.scale = TRUE, Y matrix is scaled to unit variances (default: FALSE).}

\item{remove_near_zero_variance}{Logical. If remove_near_zero_variance = TRUE, remove_near_zero_variance variables will be removed.}

\item{remove_zero_variance}{Logical. If remove_zero_variance = TRUE, remove_zero_variance variables will be removed.}

\item{toKeep.zv}{Character vector. Name of variables in X to not be deleted by (near) zero variance filtering.}

\item{remove_non_significant_models}{Logical. If remove_non_significant_models = TRUE, non-significant models are removed before computing the evaluation.}

\item{alpha}{Numeric. Cutoff for establish significant variables. Below the number are considered as significant (default: 0.05).}

\item{MIN_NVAR}{Numeric If remove_non_significant = TRUE, non-significant variables in final cox model will be removed until all variables are significant (forward selection).}

\item{MAX_NVAR}{Numeric If remove_non_significant = TRUE, non-significant variables in final cox model will be removed until all variables are significant (forward selection).}

\item{n.cut_points}{Numeric. Number of start cut points for look the optimal number of variable. 2 cut points mean start with the minimum and maximum. 3 start with minimum, maximum and middle point...(default: 3)}

\item{MIN_AUC_INCREASE}{Numeric If remove_non_significant = TRUE, non-significant variables in final cox model will be removed until all variables are significant (forward selection).}

\item{EVAL_METHOD}{Numeric. If remove_non_significant = TRUE, non-significant variables in final cox model will be removed until all variables are significant (forward selection).}

\item{w_AIC}{Numeric. Weight for AIC evaluator. All three weights must sum 1 (default: 0).}

\item{w_c.index}{Numeric. Weight for C-Index evaluator. All three weights must sum 1 (default: 0).}

\item{w_AUC}{Numeric. Weight for AUC evaluator. All three weights must sum 1 (default: 1).}

\item{times}{Numeric vector. Time points where the AUC will be evaluated. If NULL, a maximum of 15 points will be selected equally distributed.}

\item{MIN_AUC}{Numeric. Minimum AUC desire.}

\item{MIN_COMP_TO_CHECK}{Numeric. Number of penalties to check whether the AUC improves.}

\item{pred.attr}{Character. Method for average the AUC. Must be one of the following: "mean" or "median" (default: "mean").}

\item{pred.method}{Character. AUC method for evaluation. Must be one of the following: "risksetROC", "survivalROC", "cenROC", "nsROC", "smoothROCtime_C", "smoothROCtime_I" (default: "cenROC")}

\item{fast_mode}{Logical. If fast_mode = TRUE, for each run, only one fold is evaluated simultaneously. If fast_mode = FALSE, for each run, all linear predictors are computed for test observations. Once all have their linear predictors, the evaluation is perform across all the observations together (default: FALSE).}

\item{max.iter}{Maximum number of iterations for PLS convergence.}

\item{MIN_EPV}{Minimum number of Events Per Variable you want reach for the final cox model. Used to restrict the number of variables can appear in cox model. If the minimum is not meet, the model is not computed.}

\item{return_models}{Logical. Return all models computed in cross validation.}

\item{PARALLEL}{Logical. Run the cross validation with multicore option. As many cores as your total cores - 1 will be used. It could lead to higher RAM consumption.}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}

\item{seed}{Number. Seed value for perform the runs/folds divisions.}
}
\value{
Instance of class "HDcox" and model "cv.sPLS-DACOX-MixOmics".
}
\description{
plsdacox_mixOmics cross validation model
}
