% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_multiblock_Functions.R
\name{deleteZeroOrNearZeroVariance.mb}
\alias{deleteZeroOrNearZeroVariance.mb}
\title{deleteZeroOrNearZeroVariance.mb}
\usage{
deleteZeroOrNearZeroVariance.mb(
  X,
  remove_near_zero_variance = F,
  remove_zero_variance = T,
  toKeep.zv = NULL,
  freqCut = 95/5
)
}
\arguments{
\item{X}{Data.frame}

\item{remove_near_zero_variance}{Logical. Remove NEAR zero variance}

\item{remove_zero_variance}{Logical. Remove ZERO variance}

\item{toKeep.zv}{Variables to keep.}

\item{freqCut}{the cutoff for the ratio of the most common value to the second most common value}
}
\description{
deleteZeroOrNearZeroVariance.mb
}
