% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{deleteZeroOrNearZeroVariance}
\alias{deleteZeroOrNearZeroVariance}
\title{deleteZeroOrNearZeroVariance}
\usage{
deleteZeroOrNearZeroVariance(
  X,
  remove_near_zero_variance = F,
  remove_zero_variance = T,
  toKeep.zv = NULL,
  freqCut = 95/5
)
}
\arguments{
\item{X}{Numeric matrix or data.frame. Explanatory variables. Qualitative variables must be transform into binary variables.}

\item{remove_near_zero_variance}{Logical. If remove_near_zero_variance = TRUE, near zero variance variables will be removed (default: TRUE).}

\item{remove_zero_variance}{Logical. If remove_zero_variance = TRUE, zero variance variables will be removed (default: TRUE).}

\item{toKeep.zv}{Character vector. Name of variables in X to not be deleted by (near) zero variance filtering (default: NULL).}

\item{freqCut}{Numeric. Cutoff for the ratio of the most common value to the second most common value (default: 95/5).}
}
\description{
The function allows users to delete variables with near-zero or zero variance in a
given dataset. The function has parameters to control whether near-zero variance variables should
be removed or just zero variance variables. The function is based on the package and function caret::nearZeroVar().
}
\examples{
\dontrun{
# zero-variance variables
deleteZeroOrNearZeroVariance(X, remove_near_zero_variance = F, remove_zero_variance = T)
# near-zero-variance variables
deleteZeroOrNearZeroVariance(X, remove_near_zero_variance = T, remove_zero_variance = T)
}
}
