% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{deleteZeroVarianceVariables}
\alias{deleteZeroVarianceVariables}
\title{deleteZeroVarianceVariables}
\usage{
deleteZeroVarianceVariables(
  data,
  mustKeep = NULL,
  names = NULL,
  info = T,
  freqCut = 95/5,
  onlyZero = F
)
}
\arguments{
\item{data}{Numeric matrix or data.frame to remove Near Zero Variance}

\item{mustKeep}{Character vector. Variables to not remove.}

\item{names}{If data is a list. Name to access to the data.frame.}

\item{info}{Logical. If info = TRUE, list of variables deleted are return.}

\item{freqCut}{the cutoff for the ratio of the most common value to the second most common value. More info in "caret::nearZeroVar".}

\item{onlyZero}{Logical. If onlyZero = TRUE, just Zero Variance variables are removed.}
}
\value{
\itemize{
\item \code{(filteredData)}: data filtered matrix/data.frame.
\item \code{(variablesDeleted)}: If info, vector of variables deleted. Else, NULL.
}
}
\description{
deleteZeroVarianceVariables
}
