% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{eval_HDcox_model_per_variable}
\alias{eval_HDcox_model_per_variable}
\title{eval_HDcox_model_per_variable}
\usage{
eval_HDcox_model_per_variable(
  model,
  X_test,
  Y_test,
  pred.method = "cenROC",
  pred.attr = "mean",
  times = NULL,
  max_time_points = 15,
  PARALLEL = F,
  verbose = F
)
}
\arguments{
\item{model}{HDcox model.}

\item{X_test}{Numeric matrix or data.frame. Explanatory variables for test data (raw format). Qualitative variables must be transform into binary variables.}

\item{Y_test}{Numeric matrix or data.frame. Response variables for test data. Object must have two columns named as "time" and "event". For event column, accepted values are: 0/1 or FALSE/TRUE for censored and event observations.}

\item{pred.method}{Character. AUC evaluation algorithm method for evaluate the model performance. Must be one of the following: "risksetROC", "survivalROC", "cenROC", "nsROC", "smoothROCtime_C", "smoothROCtime_I" (default: "cenROC").}

\item{pred.attr}{Character. Way to evaluate the metric selected. Must be one of the following: "mean" or "median" (default: "mean").}

\item{times}{Numeric vector. Time points where the AUC will be evaluated. If NULL, a maximum of 'max_time_points' points will be selected equally distributed (default: NULL).}

\item{max_time_points}{Numeric. Maximum number of time points to use for evaluating the model (default: 15).}

\item{PARALLEL}{Logical. Run the cross validation with multicore option. As many cores as your total cores - 1 will be used. It could lead to higher RAM consumption (default: FALSE).}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}
}
\description{
The function is used to evaluate how each variable/component affect the prediction for an specific HDcox model. The function
returns all the data needed for plotting the information, including the AUC per model and for each
time point selected for the predictions. After their used, it is recommended to run the plot_evaluation()
function.
}
\examples{
\dontrun{
X <- data[train_index,]
Y <- data_Y[train_index,]
X_test <- data[-train_index,]
Y_test <- data_Y[-train_index,]
model_icox <- splsicox(X, Y)
eval_HDcox_model_per_variable(model_icox, X_test, Y_test, pred.method = "cenROC")
}
}
