% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{eval_models4.0}
\alias{eval_models4.0}
\title{eval_models4.0}
\usage{
eval_models4.0(
  lst_models,
  X_test,
  Y_test,
  pred.method,
  pred.attr = "mean",
  times = NULL,
  PARALLEL = F,
  max_time_points = 15,
  verbose = F,
  progress_bar = T
)
}
\arguments{
\item{lst_models}{List of HDcox models (must be named)}

\item{X_test}{X test}

\item{Y_test}{Y test}

\item{pred.method}{Evaluator method}

\item{pred.attr}{Mean or median}

\item{times}{Vector of time points}

\item{PARALLEL}{Use multicore option.}

\item{max_time_points}{maximum number of time points to compute}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}

\item{progress_bar}{Show progress bar.}
}
\description{
Evaluate multiple HDcox models simultaneously.
}
