% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{event.density.plot}
\alias{event.density.plot}
\title{event.density.plot}
\usage{
event.density.plot(
  Y,
  round = F,
  roundTo = 0.25,
  categories = c("Death", "Censored")
)
}
\arguments{
\item{Y}{Y matrix}

\item{round}{Logical. Round time}

\item{roundTo}{Round time to which value.}

\item{categories}{Categories to print-}
}
\description{
event.density.plot
}
