% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{getAutoKM.list}
\alias{getAutoKM.list}
\title{getAutoKM.list}
\usage{
getAutoKM.list(
  type = "LP",
  lst_models,
  comp = 1:2,
  top = 10,
  ori_data = T,
  BREAKTIME = NULL,
  n.breaks = 20,
  only_sig = F,
  alpha = 0.05,
  title = NULL,
  verbose = FALSE
)
}
\arguments{
\item{type}{Kaplan Meier for linear predictors ("LP"), for PLS components ("COMP") or for original variables ("VAR").}

\item{lst_models}{List of HDcox models}

\item{comp}{vector of two components}

\item{top}{select top X variables}

\item{ori_data}{Compute the KM with the original data cutpoints or the normalize}

\item{BREAKTIME}{Break for time point}

\item{n.breaks}{If BREAKTIME is NULL, number of time-breaks}

\item{only_sig}{Return only significant log-rank test variables}

\item{alpha}{Significant cutoff}

\item{title}{Title of the plot}

\item{verbose}{Return messages}
}
\description{
getAutoKM.list
}
