% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{getCutoffAutoKM.list}
\alias{getCutoffAutoKM.list}
\title{getCutoffAutoKM.list}
\usage{
getCutoffAutoKM.list(lst_results)
}
\arguments{
\item{lst_results}{result of getAutoKM,list().}
}
\description{
getCutoffAutoKM.list
}
