% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_multiblock_Functions.R
\name{getEPV.mb}
\alias{getEPV.mb}
\title{getEPV.mb}
\usage{
getEPV.mb(X, Y)
}
\arguments{
\item{X}{Numeric matrix or data.frame. Explanatory variables. Qualitative variables must be transform into binary variables.}

\item{Y}{Numeric matrix or data.frame. Response variables. Object must have two columns named as "time" and "event". For event column, accepted values are: 0/1 or FALSE/TRUE for censored and event observations.}
}
\description{
The function computes the (Events per Variable) EPV value for a given multiblock dataset.
The function calculates the ratio of events in the Y matrix and the number of variables in the list X
object
}
\examples{
\dontrun{
getEPV.mb(X,Y)
}
}
