% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions_others.R
\name{loadingplot.fromVector.Coxmos}
\alias{loadingplot.fromVector.Coxmos}
\title{loadingplot.fromVector.Coxmos}
\usage{
loadingplot.fromVector.Coxmos(
  model,
  vector,
  zero.rm = F,
  top = NULL,
  auto.limits = T
)
}
\arguments{
\item{model}{Coxmos model.}

\item{vector}{Vector of loading}

\item{zero.rm}{Logical. Remove variables equal to 0 (default: FALSE).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}
}
\description{
loadingplot.fromVector.Coxmos
}
