% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions_others.R
\name{loadingplot.HDcox}
\alias{loadingplot.HDcox}
\title{loadingplot.HDcox}
\usage{
loadingplot.HDcox(model, zero.rm = T, top = NULL, auto.limits = T)
}
\arguments{
\item{model}{HDcox model.}

\item{zero.rm}{Logical. Remove variables equal to 0 (default: TRUE).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}
}
\description{
loadingplot.HDcox
}
