% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_common_functions.R
\name{norm01}
\alias{norm01}
\title{norm01}
\usage{
norm01(x)
}
\arguments{
\item{x}{Numeric matrix or data.frame. Explanatory variables. Only qualitative variables will be transformed into binary variables.}
}
\description{
Normalize all values into 0-1 range.
}
