% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_cox.event.list}
\alias{plot_cox.event.list}
\title{plot_cox.event.list}
\usage{
plot_cox.event.list(lst_models, type = "lp", h.breaks = 30)
}
\arguments{
\item{lst_models}{List of HDcox models}

\item{type}{"lp", "risk", "expected" or "survival" methodology}

\item{h.breaks}{Number of time breaks.}
}
\description{
plot_cox.event.list
}
