% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_Coxmos.MB.PLS.model}
\alias{plot_Coxmos.MB.PLS.model}
\title{plot_Coxmos.MB.PLS.model}
\usage{
plot_Coxmos.MB.PLS.model(
  model,
  comp = c(1, 2),
  mode = "scores",
  factor = NULL,
  legend_title = NULL,
  top = NULL,
  only_top = F,
  radius = NULL,
  names = T,
  colorReverse = F,
  text.size = 2,
  overlaps = 10
)
}
\arguments{
\item{model}{Coxmos model.}

\item{comp}{Numeric vector. Vector of length two. Select which components to plot (default: c(1,2)).}

\item{mode}{Character. Choose one of the following plots: "scores", "loadings" o "biplot" (default: "scores").}

\item{factor}{Factor. Factor variable to color the observations. If factor = NULL, event will be used (default: NULL).}

\item{legend_title}{Character. Legend title (default: NULL).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}

\item{only_top}{Logical. If "only_top" = TRUE, then only top/radius loading variables will be shown in loading or biplot graph (default: FALSE).}

\item{radius}{Numeric. Radius size (loading/scale value) to plot variable names that are greater than the radius value (default: NULL).}

\item{names}{Logical. Show loading names for top variables or for those that are outside the radius size (default: TRUE).}

\item{colorReverse}{Logical. Reverse palette colors (default: FALSE).}

\item{text.size}{Numeric. Text size (default: 2).}

\item{overlaps}{Numeric. Number of overlaps to show when plotting loading names (default: 10).}
}
\description{
plot_Coxmos.MB.PLS.model
}
