% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_divergent.biplot}
\alias{plot_divergent.biplot}
\title{plot_divergent.biplot}
\usage{
plot_divergent.biplot(
  X,
  Y,
  NAMEVAR1,
  NAMEVAR2,
  BREAKTIME,
  x.text = "N. of Samples"
)
}
\arguments{
\item{X}{Numeric matrix or data.frame. Explanatory variables with "NAMEVAR1" and "NAMEVAR2" variables. "NAMEVAR1" must be a factor variable.}

\item{Y}{Numeric matrix or data.frame. Response variables. Object must have two columns named as "time" and "event". For event column, accepted values are: 0/1 or FALSE/TRUE for censored and event observations.}

\item{NAMEVAR1}{Character. Factor variable name (must be located in colnames(X) and have to have two levels).}

\item{NAMEVAR2}{Character. Numerical variable name (must be located in colnames(X)).}

\item{BREAKTIME}{Numeric. Size of time to split the data into "total_time / BREAKTIME + 1" points. If BREAKTIME = NULL, "n.breaks" is used (default: NULL).}

\item{x.text}{Character. Title for X axis.}
}
\value{
A ggplot2 two side bar plot. X axis represent the number of samples per each NAMEVAR1 factor levels and Y axis, the X NAMEVAR2 numerical variables categorize in groups of breaks.
}
\description{
Two side plot by a qualitative and a quantitative variable and Y event matrix.

Generates a two-sided plot that visualizes the relationship between a qualitative
variable and a quantitative variable, along with the event/censoring status in the data.
The qualitative variable is used to split the data into two sides, and then the quantitative variable
is used to divide the data into different sections or intervals, which are displayed as horizontal bars
on each side of the plot. The left side of the plot shows the data for one level of the qualitative
variable, while the right side shows the data for the other level. Each side is colored according
to the event/censored quantity of observations. The parameters "X" and "Y" are the data frames containing
the variables to be plotted, "NAMEVAR1" and "NAMEVAR2" are the names of the qualitative and
quantitative variables, respectively.
}
\details{
For time comparison between best models, use the cross validation objects instead the final models. The training has to be taken into account.
}
\examples{
\dontrun{
  NAMEVAR1 = "sex"
  NAMEVAR2 = "age"
  plot_divergent.biplot(X, Y, NAMEVAR1, NAMEVAR2, BREAKTIME = 5, x.text = "N. of Patients")
}
}
