% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_evaluation.list}
\alias{plot_evaluation.list}
\title{plot_evaluation.list}
\usage{
plot_evaluation.list(
  lst_eval_results,
  pred.attr = "mean",
  y.min = NULL,
  type = "both"
)
}
\arguments{
\item{lst_eval_results}{List (named) of HDcox evaluation results.}

\item{pred.attr}{Character. Way to evaluate the metric selected. Must be one of the following: "mean" or "median" (default: "mean").}

\item{y.min}{Numeric. Minimum Y value for establish the Y axis value. If y.min = NULL, automatic detection is performed (default: NULL).}

\item{type}{Character. Plot type. Must be one of the following: "both", "line" or "mean". In other case, "both" will be selected (default: "both").}
}
\description{
Run the function "plot_evaluation" for a list of results. More information in "?plot_evaluation".
}
\examples{
\dontrun{
#lst_eval_results #list of multiple eval_HDcox_models()
plot_evaluation.list(lst_eval_results)
}
}
