% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_evaluation}
\alias{plot_evaluation}
\title{plot_evaluation}
\usage{
plot_evaluation(eval_results, pred.attr = "mean", y.min = NULL, type = "both")
}
\arguments{
\item{eval_results}{Eval_models4.0 object}

\item{pred.attr}{"mean" or "median"}

\item{y.min}{Minimum Y value to plot. If NULL, automatic detection.}

\item{type}{type plot. Must be one of the following: "both", "line", "mean". In other case, "both" will be selected.}
}
\description{
plot_evaluation
}
