% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_events}
\alias{plot_events}
\title{plot_events}
\usage{
plot_events(
  Y,
  roundTo = 0.25,
  categories = c("Censored", "Death"),
  y.text = "Number of patients",
  max.breaks = 20
)
}
\arguments{
\item{Y}{Y matrix}

\item{roundTo}{Round time to which value.}

\item{categories}{Categories to print.}

\item{y.text}{Label for Y axis}

\item{max.breaks}{Number maximum of breaks in X axis.}
}
\description{
plot_events
}
