% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_HDcox.PLS.model}
\alias{plot_HDcox.PLS.model}
\title{plotHDcox.PLS}
\usage{
plot_HDcox.PLS.model(
  model,
  comp = c(1, 2),
  mode = "scores",
  factor = NULL,
  legend.title = NULL,
  mahalanovis_limit = 20,
  radius = 0.2,
  names = F,
  allNames = F,
  colorReverse = F,
  text.size = 4
)
}
\arguments{
\item{model}{HDcox model}

\item{comp}{vector of two components}

\item{mode}{scores, loadings o biplot}

\item{factor}{factor to color}

\item{legend.title}{Legend title}

\item{mahalanovis_limit}{Mahalanobis distance limit}

\item{radius}{Radious to plot variable names}

\item{names}{Logical. Show names for variables outside the radius.}

\item{allNames}{Logical. Show all variable names.}

\item{colorReverse}{Logical. Color reverse.}

\item{text.size}{Text size.}
}
\description{
plotHDcox.PLS
}
