% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_patient.eventDensity}
\alias{plot_patient.eventDensity}
\title{patient.eventDensity}
\usage{
plot_patient.eventDensity(
  patient,
  time = NULL,
  model,
  type = "lp",
  size = 3,
  color = "red"
)
}
\arguments{
\item{patient}{observation}

\item{time}{time point of study}

\item{model}{HDcox model}

\item{type}{type between "lp", "risk, "expected", "survival"}

\item{size}{size}

\item{color}{color}
}
\description{
patient.eventDensity
}
