% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{plot_patient.eventHistogram}
\alias{plot_patient.eventHistogram}
\title{patient.eventHistogram}
\usage{
plot_patient.eventHistogram(
  patient,
  time = NULL,
  model,
  type = "lp",
  size = 3,
  color = "red"
)
}
\arguments{
\item{patient}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation. Qualitative variables must be transform into binary variables.}

\item{time}{Numeric. Time point where the AUC will be evaluated (default: NULL).}

\item{model}{Coxmos model.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{size}{Numeric. Point size (default: 3).}

\item{color}{String. R Color.}
}
\description{
Generates a histogram plot for patient event data based on a given Coxmos model. The function visualizes the distribution of predicted values and highlights the prediction for a specific patient.
}
\details{
The \code{plot_patient.eventHistogram} function is designed to provide a visual representation of the distribution of predicted event values based on a Coxmos model. The function takes in patient data, a specified time point, and a Coxmos model to compute the prediction. The resulting histogram plot displays the distribution of these predictions, with a specific emphasis on the prediction for the provided patient data. The prediction is represented as a point on the histogram, allowing for easy comparison between the specific patient's prediction and the overall distribution of predictions. The type of prediction ("lp", "risk", "expected", or "survival") can be specified, offering flexibility in the kind of insights one wishes to derive from the visualization. The appearance of the point representing the patient's prediction can be customized using the \code{size} and \code{color} parameters.
}
\examples{
\dontrun{
plot_patient.eventHistogram(patient, time = NULL, model)
}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
