% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{plot_proportionalHazard.list}
\alias{plot_proportionalHazard.list}
\title{plot_proportionalHazard.list}
\usage{
plot_proportionalHazard.list(lst_models)
}
\arguments{
\item{lst_models}{List of Coxmos models.}
}
\description{
Run the function "plot_proportionalHazard" for a list of models. More information in
"?plot_proportionalHazard".
}
\examples{
\dontrun{
plot_proportionalHazard.list(lst_models)
}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
