% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_proportionalHazard}
\alias{plot_proportionalHazard}
\title{plot_proportionalHazard}
\usage{
plot_proportionalHazard(model)
}
\arguments{
\item{model}{Coxmos model.}
}
\value{
A ggplot2 plot.
}
\description{
Perform the proportional hazard assumption for a specific Coxmos model.
It uses the functions "survival::cox.zph" and "survminer::ggcoxzph", but it returns a ggplot2 graph.
}
\examples{
\dontrun{
plot_proportionalHazard(model)
}
}
