% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta.list}
\alias{plot_pseudobeta.list}
\title{plot_pseudobeta.list}
\usage{
plot_pseudobeta.list(
  lst_models,
  error.bar = T,
  onlySig = F,
  alpha = 0.05,
  zero.rm = F,
  top = NULL,
  auto.limits = T
)
}
\arguments{
\item{lst_models}{List of HDcox model}

\item{error.bar}{Logical. Show error bar}

\item{onlySig}{Logical. Compute psudobetas using only significant components.}

\item{alpha}{Significant value (P.value).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0.}

\item{top}{Plot the top X variables with the higher pseudobetas in absolute value.}

\item{auto.limits}{Compute Y limit automatically}
}
\description{
plot_pseudobeta.list
}
