% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta}
\alias{plot_pseudobeta}
\title{plot_pseudobeta}
\usage{
plot_pseudobeta(
  model,
  error.bar = T,
  onlySig = F,
  alpha = 0.05,
  zero.rm = T,
  top = NULL,
  auto.limits = T,
  show_percentage = T,
  size_percentage = 3
)
}
\arguments{
\item{model}{HDcox model.}

\item{error.bar}{Logical. Show error bar (default: TRUE).}

\item{onlySig}{Logical. Compute psudobetas using only significant components (default: FALSE).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0 (default: TRUE).}

\item{top}{Numeric. Show "top" first variables with the higher pseudobetas in absolute value. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{show_percentage}{Logical. If show_percentage = TRUE, it shows the contribution percentage for each variable to the full model (default: TRUE).}

\item{size_percentage}{Numeric. Size of percentage text (default: 3).}
}
\description{
This function takes a HDcox PLS model and decomposes it into an original variable
pseudo-beta interpretation. It expresses the Cox final formula in terms of original variables,
using weights associated with each component. The function can also show the percentage contribution
of each variable to the final model, either for all variables or for a specified number of top
variables. Additionally, error bars can be added and only significant variables can be displayed.
}
\examples{
\dontrun{
plot_pseudobeta(model)
}
}
