% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta_newPatient.list}
\alias{plot_pseudobeta_newPatient.list}
\title{plot_pseudobeta_newPatient.list}
\usage{
plot_pseudobeta_newPatient.list(
  lst_models,
  new_observation,
  error.bar = T,
  onlySig = T,
  alpha = 0.05,
  zero.rm = T,
  top = NULL,
  auto.limits = T,
  show.betas = F,
  verbose = F
)
}
\arguments{
\item{lst_models}{List of HDcox models.}

\item{new_observation}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation. Qualitative variables must be transform into binary variables.}

\item{error.bar}{Logical. Show error bar (default: TRUE).}

\item{onlySig}{Logical. Compute psudobetas using only significant components (default: TRUE).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0 (default: TRUE).}

\item{top}{Numeric. Show "top" first variables with the higher pseudobetas in absolute value. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{show.betas}{Logical. Show original betas (default: FALSE).}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}
}
\description{
Run the function "plot_pseudobeta_newPatient" for a list of models. More information in "?plot_pseudobeta_newPatient".
}
\examples{
\dontrun{
plot_pseudobeta_newPatient.list(lst_models, new_observation)
}
}
