% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_time.models}
\alias{plot_time.models}
\title{Time consuming plot.}
\usage{
plot_time.models(lst_models, x.text = "Method", y.text = NULL)
}
\arguments{
\item{lst_models}{A list of HDcox objects. Each HDCox object has the attribute time measured in minutes.}

\item{x.text}{X axis title.}

\item{y.text}{Y axis title. (default: NULL == "Time (mins)").}
}
\value{
A ggplot2 bar plot for each object in lst_models and the total of all of them.
}
\description{
Time consuming plot.
}
\details{
For time comparison between best models, use the cross validation objects instead the final models. The training has to be taken into account.
}
\examples{
\dontrun{
  lst_models = {"cox" = cox_model, "PLS-ICOX" = cv.plsicox_model, "sPLS-DRCOX" = cv.splsdrcox_model}
  plot_time.models(lst_models, x.text = "Method")
}
}
