% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{predict.HDcox}
\alias{predict.HDcox}
\title{predict.HDcox}
\usage{
\method{predict}{HDcox}(object, ..., newdata = NULL)
}
\arguments{
\item{object}{HDcox model}

\item{...}{additional arguments affecting the predictions produced.}

\item{newdata}{Numeric matrix or data.frame. New data for explanatory variables (raw data). Qualitative variables must be transform into binary variables.}
}
\value{
Score values for new data using the HDcox model selected.
}
\description{
The function is used to generate the prediction score matrix for PLS Survival models.
The score matrix is required for the final survival prediction of the data, as dimensional reduction
is performed before running the cox survival analysis.
}
\examples{
\dontrun{
X <- data[train_index,]
Y <- data_Y[train_index,]
X_test <- data[-train_index,]
model <- splsicox(X, Y)
predict.HDcox(object = model, newdata = X_test)
}
}
