% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{print.HDcox}
\alias{print.HDcox}
\title{print.HDcox}
\usage{
\method{print}{HDcox}(x, ...)
}
\arguments{
\item{x}{HDcox object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print.HDcox
}
