% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{save_ggplot_lst.svg}
\alias{save_ggplot_lst.svg}
\title{save_ggplot_lst.svg}
\usage{
save_ggplot_lst.svg(
  lst_plots,
  folder = NULL,
  prefix = NULL,
  suffix = NULL,
  wide = T,
  quality = "4K",
  dpi = 80,
  custom = NULL,
  object_name = NULL
)
}
\arguments{
\item{lst_plots}{list of ggplot2}

\item{folder}{folder}

\item{prefix}{prefix for file name}

\item{suffix}{sufix for file name}

\item{wide}{16:9 or 4:3}

\item{quality}{one of: "HD", "FHD", "2K", "4K", "8K"}

\item{dpi}{dpi}

\item{custom}{Custom size. Numeric vector of width and height}

\item{object_name}{file name}
}
\description{
Allows to save a list of ggplot2 objects in .svg format based on an specific resolution.
}
\examples{
\dontrun{
save_ggplot_lst.svg(plot_lst, folder)
}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
