% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_comparison_functions.R
\name{super.evalResults2DataFrame}
\alias{super.evalResults2DataFrame}
\title{super.evalResults2DataFrame}
\usage{
super.evalResults2DataFrame(eval_results)
}
\arguments{
\item{eval_results}{Eval results}
}
\description{
super.evalResults2DataFrame
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
