% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_comparison_functions.R
\name{super.trainAllModels}
\alias{super.trainAllModels}
\title{super.trainAllModels}
\usage{
super.trainAllModels(
  lst_subdata,
  methods,
  comp_calculation = "manual",
  ncomp = 5,
  EN.alpha = 0.5,
  eta = 0.5,
  x.center = T,
  x.scale = T,
  y.center = F,
  y.scale = F,
  MIN_EPV = 0,
  PARALLEL = T
)
}
\arguments{
\item{lst_subdata}{List of data (X, Y)}

\item{methods}{Methods to use for train}

\item{comp_calculation}{Compute optimal number of components by "manual" or "auto".}

\item{ncomp}{N.components for manual detection.}

\item{EN.alpha}{Penalization for manual detection,}

\item{eta}{Eta for manual detection,}

\item{x.center}{Logical. If x.center = TRUE, X matrix is centered to zero means (default: TRUE).}

\item{x.scale}{Logical. If x.scale = TRUE, X matrix is scaled to unit variances (default: FALSE).}

\item{y.center}{Logical. If y.center = TRUE, Y matrix is centered to zero means (default: FALSE).}

\item{y.scale}{Logical. If y.scale = TRUE, Y matrix is scaled to unit variances (default: FALSE).}

\item{MIN_EPV}{Minimum number of Events Per Variable you want reach for the final cox model. Used to restrict the number of variables can appear in cox model. If the minimum is not meet, the model is not computed.}

\item{PARALLEL}{Logical. Run the cross validation with multicore option. As many cores as your total cores - 1 will be used. It could lead to higher RAM consumption.

@export}
}
\description{
super.trainAllModels
}
