% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{w.starplot.HDcox}
\alias{w.starplot.HDcox}
\title{w.starplot.HDcox}
\usage{
w.starplot.HDcox(model, zero.rm = F, top = NULL, auto.limits = T)
}
\arguments{
\item{model}{HDcox model}

\item{zero.rm}{Remove variables equal to 0.}

\item{top}{Number. Show top variables.}

\item{auto.limits}{Logical. If TRUE, limits are detected for a better plot.}
}
\description{
w.starplot.HDcox
}
