% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_miRNA_glioblastoma}
\alias{X_miRNA_glioblastoma}
\title{X_miRNA_glioblastoma Data}
\format{
A data frame with 552 observations and 534 features:
\describe{miRNA data set from multiomic glioblastome data.}
}
\source{
Glioblastome data
}
\usage{
X_miRNA_glioblastoma
}
\description{
miRNA dataset from Glioblastome multiomic dataset
}
\keyword{datasets}
