% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_multiomic}
\alias{X_multiomic}
\title{X_multiomic Data}
\format{
A data frame with 150 observations and two omics (miRNA and proteomic):
\describe{642 miRNAs, 369 proteins}
}
\source{
TCGA-BRCA data
}
\usage{
X_multiomic
}
\description{
Toy dataset from BREAST CANCER. miRNA and Protein data. (https://github.com/pilargmarch/multiomics2.0/tree/main)
}
\keyword{datasets}
