% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_proteomic}
\alias{X_proteomic}
\title{X_proteomic Data}
\format{
A data frame with 150 observations and 369 features:
\describe{Small data set from original data (585 observations).}
}
\source{
TCGA-BRCA data
}
\usage{
X_proteomic
}
\description{
Toy dataset from BREAST CANCER. Protein data. (https://github.com/pilargmarch/multiomics2.0/tree/main)
}
\keyword{datasets}
