% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_small_data_E.MTAB.386}
\alias{X_small_data_E.MTAB.386}
\title{X_small_data_E.MTAB.386 Data}
\format{
A data frame with 126 observations and 500 features:
\describe{Each of the variables is a random gene selected from the original dataset}
}
\source{
curatedOvarianData R package
}
\usage{
X_small_data_E.MTAB.386
}
\description{
Toy dataset from E.MTAB.386 ('curatedOvarianData' R package)
}
\keyword{datasets}
