% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_small_mo.data_Glioblastoma}
\alias{X_small_mo.data_Glioblastoma}
\title{X_small_mo.data_Glioblastoma Data}
\format{
A data frame with 225 observations and three block of features:
\describe{500 genes, 500 miRNA and 37 clinical features}
}
\source{
!!!
}
\usage{
X_small_mo.data_Glioblastoma
}
\description{
Toy dataset from Glioblastoma multiomic data (!!!)
}
\keyword{datasets}
