% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_miRNA_glioblastoma}
\alias{Y_miRNA_glioblastoma}
\title{Y_miRNA_glioblastoma Data}
\format{
A data frame with 552 observations and 2 features:
\describe{
\item{time}{Global survival time in years. Time to the event of to the last patient information.}
\item{event}{Numeric. FALSE for censored and TRUE for event observations.}
}
}
\source{
Glioblastome data
}
\usage{
Y_miRNA_glioblastoma
}
\description{
miRNA dataset from Glioblastome multiomic dataset
}
\keyword{datasets}
