% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_multiomic}
\alias{Y_multiomic}
\title{Y_multiomic Data}
\format{
A data frame with 150 observations and 2 features:
\describe{
\item{time}{Global survival time in years. Time to the event of to the last patient information.}
\item{event}{Numeric. FALSE/0 for censored and TRUE/1 for event observations.}
}
}
\source{
TCGA-BRCA data
}
\usage{
Y_multiomic
}
\description{
Toy dataset from BREAST CANCER. miRNA and Protein data. (https://github.com/pilargmarch/multiomics2.0/tree/main)
}
\keyword{datasets}
