% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_small_data_E.MTAB.386}
\alias{Y_small_data_E.MTAB.386}
\title{Y_small_data_E.MTAB.386 Data}
\format{
A data frame with 126 observations and 2 features:
\describe{
\item{time}{Global survival time in years. Time to the event of to the last patient information.}
\item{event}{Numeric. 0 for censored and 1 for event observations.}
}
}
\source{
curatedOvarianData R package
}
\usage{
Y_small_data_E.MTAB.386
}
\description{
Toy dataset from E.MTAB.386 ('curatedOvarianData' R package)
}
\keyword{datasets}
