% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{cox.prediction}
\alias{cox.prediction}
\title{cox.prediction}
\usage{
cox.prediction(model, new_data, time = NULL, type = "lp", method = "cox")
}
\arguments{
\item{model}{Coxmos model.}

\item{new_data}{Numeric matrix or data.frame. New explanatory variables (raw data). Qualitative variables must be transform into binary variables.}

\item{time}{Numeric. Time point where the AUC will be evaluated (default: NULL).}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{method}{Character. Prediction method. It can be compute by using the cox model "cox" or by using W.star "W.star" (default: "cox"). (not implemented for MB approaches)!!!}
}
\value{
Return the lp or other metric for the patient
}
\description{
The function is used to perform a predict.cox() for a Coxmos model from raw data. The function
automatically generates the score matrix if a PLS Survival is performed and then performs the cox prediction.
}
\examples{
\dontrun{
X <- data[train_index,]
X_test <- data[-train_index,]
model <- plsicox(X, Y)
cox.prediction(model = model, new_data = X_test)
}
}
