% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_multiblock_Functions.R
\name{deleteNearZeroCoefficientOfVariation.mb}
\alias{deleteNearZeroCoefficientOfVariation.mb}
\title{deleteNearZeroCoefficientOfVariation.mb}
\usage{
deleteNearZeroCoefficientOfVariation.mb(X, LIMIT = 0.1)
}
\arguments{
\item{X}{List of numeric matrices or data.frames. Explanatory variables. Qualitative variables must be transform into binary variables.}

\item{LIMIT}{Numeric. Cutoff for minimum variation. If coefficient is lesser than the limit, the variables are removed because not vary enough (default: 0.1).}
}
\description{
deleteNearZeroCoefficientOfVariation.mb
}
