% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_common_functions.R
\name{deleteNearZeroCoefficientOfVariation}
\alias{deleteNearZeroCoefficientOfVariation}
\title{deleteNearZeroCoefficientOfVariation}
\usage{
deleteNearZeroCoefficientOfVariation(X, LIMIT = 0.1)
}
\arguments{
\item{X}{Numeric matrix or data.frame. Explanatory variables. Qualitative variables must be transform into binary variables.}

\item{LIMIT}{Numeric. Cutoff for minimum variation. If coefficient is lesser than the limit, the variables are removed because not vary enough (default: 0.1).}
}
\description{
deleteNearZeroCoefficientOfVariation
}
