% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_multiblock_Functions.R
\name{deleteZeroOrNearZeroVariance.mb}
\alias{deleteZeroOrNearZeroVariance.mb}
\title{deleteZeroOrNearZeroVariance.mb}
\usage{
deleteZeroOrNearZeroVariance.mb(
  X,
  remove_near_zero_variance = F,
  remove_zero_variance = T,
  toKeep.zv = NULL,
  freqCut = 95/5
)
}
\arguments{
\item{X}{List of numeric matrices or data.frame. Explanatory variables. Qualitative variables must be transform into binary variables.}

\item{remove_near_zero_variance}{Logical. If remove_near_zero_variance = TRUE, near zero variance variables will be removed (default: TRUE).}

\item{remove_zero_variance}{Logical. If remove_zero_variance = TRUE, zero variance variables will be removed (default: TRUE).}

\item{toKeep.zv}{Character vector. Name of variables in X to not be deleted by (near) zero variance filtering (default: NULL).}

\item{freqCut}{Numeric. Cutoff for the ratio of the most common value to the second most common value (default: 95/5).}
}
\description{
deleteZeroOrNearZeroVariance.mb
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
