% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{deleteZeroOrNearZeroVariance}
\alias{deleteZeroOrNearZeroVariance}
\title{deleteZeroOrNearZeroVariance}
\usage{
deleteZeroOrNearZeroVariance(
  X,
  remove_near_zero_variance = F,
  remove_zero_variance = T,
  toKeep.zv = NULL,
  freqCut = 95/5
)
}
\arguments{
\item{X}{Data.frame}

\item{remove_near_zero_variance}{Logical. Remove NEAR zero variance}

\item{remove_zero_variance}{Logical. Remove ZERO variance}

\item{toKeep.zv}{Variables to keep.}

\item{freqCut}{the cutoff for the ratio of the most common value to the second most common value}
}
\description{
deleteZeroOrNearZeroVariance
}
