% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{getAutoKM}
\alias{getAutoKM}
\title{getAutoKM}
\usage{
getAutoKM(
  type = "LP",
  model,
  comp = 1:2,
  top = 10,
  ori_data = T,
  BREAKTIME = NULL,
  n.breaks = 20,
  only_sig = F,
  alpha = 0.05,
  title = NULL,
  verbose = FALSE
)
}
\arguments{
\item{type}{Character. Kaplan Meier for complete model linear predictor ("LP"), for PLS components ("COMP") or for original variables ("VAR") (default: LP).}

\item{model}{Coxmos model.}

\item{comp}{Numeric vector. Vector of length two. Select which components to plot (default: c(1,2)).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: 10).}

\item{ori_data}{Logical. Compute the Kaplan-Meier plot with the raw-data or the normalize-data to compute the best cut-point for splitting the data into two groups. Only used when type = "VAR" (default: TRUE).}

\item{BREAKTIME}{Numeric. Size of time to split the data into "total_time / BREAKTIME + 1" points. If BREAKTIME = NULL, "n.breaks" is used (default: NULL).}

\item{n.breaks}{Numeric. If BREAKTIME is NULL, "n.breaks" is the number of time-break points to compute (default: 20).}

\item{only_sig}{Logical. If "only_sig" = TRUE, then only significant log-rank test variables are returned (default: FALSE).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{title}{Character. Kaplan-Meier plot title (default: NULL).}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}
}
\description{
Performs a Kaplan-Meier plot for the Coxmos model selected. KM could be based on model Linear Predictor value ("LP"), based on PLS-COX component ("COMP") or at original variable level ("VAR").
}
\examples{
\dontrun{
getAutoKM(type = "LP", model)
}
}
\references{
\insertRef{Kaplan_1958}{Coxmos}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
