% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{getCutoffAutoKM.list}
\alias{getCutoffAutoKM.list}
\title{getCutoffAutoKM.list}
\usage{
getCutoffAutoKM.list(lst_results)
}
\arguments{
\item{lst_results}{List of lists. Result of getAutoKM.list() function.}
}
\description{
Run the function "getCutoffAutoKM" for a list of models. More information in "?getCutoffAutoKM".
}
\examples{
\dontrun{
getCutoffAutoKM.list(lst_results)
}
}
\references{
\insertRef{Kaplan_1958}{Coxmos}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
