% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{getCutoffAutoKM}
\alias{getCutoffAutoKM}
\title{getCutoffAutoKM}
\usage{
getCutoffAutoKM(result)
}
\arguments{
\item{result}{List. Result of getAutoKM() function.}
}
\description{
Gets the cutoff value from the results of getAutoKM() functions.
}
\examples{
\dontrun{
getCutoffAutoKM(result)
}
}
\references{
\insertRef{Kaplan_1958}{Coxmos}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
