% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{getCutoffAutoKM}
\alias{getCutoffAutoKM}
\title{getCutoffAutoKM}
\usage{
getCutoffAutoKM(result)
}
\arguments{
\item{result}{List. Result of getAutoKM() function.}
}
\description{
This function takes the results of the getAutoKM() function and extracts the cutoff
value that was determined for the Kaplan-Meier plot. The cutoff value is the value of the predictor
variable at which the data is split into high-risk and low-risk groups.
}
\examples{
\dontrun{
getCutoffAutoKM(result)
}
}
