% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{getTestKM}
\alias{getTestKM}
\title{getTestKM}
\usage{
getTestKM(
  model,
  X_test,
  Y_test,
  cutoff,
  type = "LP",
  ori_data = T,
  BREAKTIME = NULL,
  n.breaks = 20,
  title = NULL
)
}
\arguments{
\item{model}{HDcox model}

\item{X_test}{X_test matrix (not normalized)}

\item{Y_test}{Y_test matrix}

\item{cutoff}{Cutoff to divide the observations. Recommended to compute optimal cutoff value with getAutoKM().}

\item{type}{Kaplan Meier for linear predictors ("LP"), for PLS components ("COMP") or for original variables ("VAR").}

\item{ori_data}{Compute the KM with the original data cutpoints or the normalize (only for 'VAR' type)}

\item{BREAKTIME}{Break for time points}

\item{n.breaks}{If BREAKTIME is NULL, number of time-breaks}

\item{title}{Kaplan-Meier plot title}
}
\description{
getTestKM
}
