% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions_others.R
\name{loadingplot.Coxmos}
\alias{loadingplot.Coxmos}
\title{loadingplot.Coxmos}
\usage{
loadingplot.Coxmos(model, zero.rm = TRUE, top = NULL, auto.limits = TRUE)
}
\arguments{
\item{model}{Coxmos model.}

\item{zero.rm}{Logical. Remove variables equal to 0 (default: TRUE).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}
}
\description{
loadingplot.Coxmos
}
