% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{loadingplot.fromVector.HDcox}
\alias{loadingplot.fromVector.HDcox}
\title{loadingplot.fromVector.HDcox}
\usage{
loadingplot.fromVector.HDcox(
  model,
  vector,
  zero.rm = F,
  top = NULL,
  auto.limits = T
)
}
\arguments{
\item{model}{HDcox model}

\item{vector}{Vector of loading}

\item{zero.rm}{Remove variables equal to 0.}

\item{top}{Number. Show top variables.}

\item{auto.limits}{Logical. If TRUE, limits are detected for a better plot.}
}
\description{
loadingplot.fromVector.HDcox
}
