% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_cox.event.list}
\alias{plot_cox.event.list}
\title{plot_cox.event.list}
\usage{
plot_cox.event.list(lst_models, type = "lp", n.breaks = 20)
}
\arguments{
\item{lst_models}{List of HDcox models.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{n.breaks}{Numeric. Number of time-break points to compute (default: 20).}
}
\description{
Run the function "plot_cox.event" for a list of models. More information in "?plot_cox.event".
}
\examples{
\dontrun{
plot_cox.event.list(lst_models)
}
}
