% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_cox.event}
\alias{plot_cox.event}
\title{plot.cox.event}
\usage{
plot_cox.event(model, type = "lp", h.breaks = 30)
}
\arguments{
\item{model}{HDcox model}

\item{type}{"lp", "risk", "expected" or "survival" methodology}

\item{h.breaks}{Number of time breaks.}
}
\description{
plot.cox.event
}
