% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_cox.event}
\alias{plot_cox.event}
\title{plot_cox.event}
\usage{
plot_cox.event(model, type = "lp", n.breaks = 20)
}
\arguments{
\item{model}{HDcox model.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{n.breaks}{Numeric. If BREAKTIME is NULL, "n.breaks" is the number of time-break points to compute (default: 20).}
}
\description{
Generates a histogram and a density plot for the HDcox model train.
The type of Cox predictions to be used can be specified using the "type" parameter, which accepts
"lp" (linear predictor), "risk", "expected", or "survival". The resulting plots are colored by
censored/event observations.
}
\examples{
\dontrun{
plot_cox.event(model)
}
}
