% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_divergent.biplot}
\alias{plot_divergent.biplot}
\title{plot_divergent.biplot}
\usage{
plot_divergent.biplot(
  X,
  Y,
  NAMEVAR1,
  NAMEVAR2,
  breaks,
  x.text = "N. of Samples"
)
}
\arguments{
\item{X}{A list of HDcox objects. Each HDCox object has the attribute time measured in minutes.}

\item{Y}{X axis title.}

\item{NAMEVAR1}{factor variable name (must be located in colnames(X)).}

\item{NAMEVAR2}{numerical variable name (must be located in colnames(X)).}

\item{breaks}{Number of values for X numerical variable per each Y axis value. Example: breaks = 5 -> Y axis values: "0-4, 5-9, 10-14..."}

\item{x.text}{Text for X axis.}
}
\value{
A ggplot2 two side bar plot. X axis represent the number of samples per each NAMEVAR1 factor levels and Y axis, the X NAMEVAR2 numerical variables categorize in groups of breaks.
}
\description{
Two side plot by a qualitative and a quantitative variable and Y event matrix.
}
\details{
For time comparison between best models, use the cross validation objects instead the final models. The training has to be taken into account.
}
\examples{
\dontrun{
  NAMEVAR1 = "sex"
  NAMEVAR2 = "age"
  plot_divergent.biplot(X, Y, NAMEVAR1, NAMEVAR2, breaks = 5, x.text = "N. of Patients")
}
}
