% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_evaluation.list}
\alias{plot_evaluation.list}
\title{plot.evaluation.list}
\usage{
plot_evaluation.list(
  eval_results,
  pred.attr = "mean",
  y.min = NULL,
  type = "both"
)
}
\arguments{
\item{eval_results}{List of eval_models4.0 results.}

\item{pred.attr}{"mean" or "median"}

\item{y.min}{Minimum Y value to plot. If NULL, automatic detection.}

\item{type}{type plot. Must be one of the following: "both", "line", "mean". In other case, "both" will be selected.}
}
\description{
plot.evaluation.list
}
