% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_evaluation}
\alias{plot_evaluation}
\title{plot_evaluation}
\usage{
plot_evaluation(eval_results, pred.attr = "mean", y.min = NULL, type = "both")
}
\arguments{
\item{eval_results}{HDcox evaluation object.}

\item{pred.attr}{Character. Way to evaluate the metric selected. Must be one of the following: "mean" or "median" (default: "mean").}

\item{y.min}{Numeric. Minimum Y value for establish the Y axis value. If y.min = NULL, automatic detection is performed (default: NULL).}

\item{type}{Character. Plot type. Must be one of the following: "both", "line" or "mean". In other case, "both" will be selected (default: "both").}
}
\description{
Performs a comparative evaluation test on the result object provided, which includes
tests for comparing the different models using "t.test", "anova", "wilcoxon", and "kruskal-wallis".
The function also generates boxplots for comparing the results using the specified test and a line
plot for each model, with the predicted time points on the x-axis and the AUC on the y-axis.
}
\examples{
\dontrun{
eval_results <- eval_HDcox_models(lst_models, X_test, Y_test, pred.method)
plot_evaluation(eval_results)
}
}
