% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_forest}
\alias{plot_forest}
\title{plot_forest}
\usage{
plot_forest(
  model,
  title = "Hazard Ratio",
  cpositions = c(0.02, 0.22, 0.4),
  fontsize = 0.7,
  refLabel = "reference",
  noDigits = 2
)
}
\arguments{
\item{model}{HDcox model.}

\item{title}{Character. Forest plot title (default: "Hazard Ratio").}

\item{cpositions}{Numeric vector. Relative positions of first three columns in the OX scale (default: c(0.02, 0.22, 0.4)).}

\item{fontsize}{Numeric. Elative size of annotations in the plot (default: 0.7).}

\item{refLabel}{Character. Label for reference levels of factor variables (default: "reference").}

\item{noDigits}{Numeric. Number of digits for estimates and p-values in the plot (default: 2).}
}
\description{
Forest plot for HDcox models using the R library plot_forest. For more information check ?survminer::ggforest.
}
