% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_HDcox.PLS.model}
\alias{plot_HDcox.PLS.model}
\title{plot_HDcox.PLS.model}
\usage{
plot_HDcox.PLS.model(
  model,
  comp = c(1, 2),
  mode = "scores",
  factor = NULL,
  legend.title = NULL,
  top = NULL,
  only_top = F,
  radius = NULL,
  names = T,
  colorReverse = F,
  text.size = 2,
  overlaps = 10
)
}
\arguments{
\item{model}{HDcox model}

\item{comp}{Numeric vector. Two values for choosing with two components to plot}

\item{mode}{scores, loadings o biplot}

\item{factor}{factor to color. If NULL, event will be used.}

\item{legend.title}{Legend title}

\item{top}{Number. How many loading names to show.}

\item{only_top}{Logical. If TRUE, only top/radius loading will be shown in loading or biplot graph}

\item{radius}{Radius to plot variable names}

\item{names}{Logical. Show loading names for top variables or outside radius}

\item{colorReverse}{Logical. Color reverse.}

\item{text.size}{Text size.}

\item{overlaps}{Number of overlaps to show loading names (default = 10)}
}
\description{
plot_HDcox.PLS.model
}
