% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_LP.multiplePatients}
\alias{plot_LP.multiplePatients}
\title{plot_LP.multiplePatients}
\usage{
plot_LP.multiplePatients(
  model,
  new_data,
  error.bar = F,
  onlySig = T,
  alpha = 0.05,
  zero.rm = T,
  auto.limits = T,
  top = NULL
)
}
\arguments{
\item{model}{Coxmos model.}

\item{new_data}{Numeric matrix or data.frame. New explanatory variables (raw data). Qualitative variables must be transform into binary variables.}

\item{error.bar}{Logical. Show error bar (default: FALSE).}

\item{onlySig}{Logical. Compute plot using only significant components (default: TRUE).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables equal to 0 (default: TRUE).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}
}
\description{
plot_LP.multiplePatients
}
\examples{
\dontrun{
plot_LP.multiplePatients(model, new_data)
}
}
