% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{plot_patient.eventDensity}
\alias{plot_patient.eventDensity}
\title{patient.eventDensity}
\usage{
plot_patient.eventDensity(
  patient,
  time = NULL,
  model,
  type = "lp",
  size = 3,
  color = "red"
)
}
\arguments{
\item{patient}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation.
Qualitative variables must be transform into binary variables.}

\item{time}{Numeric. Time point where the AUC will be evaluated (default: NULL).}

\item{model}{Coxmos model.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{size}{Numeric. Point size (default: 3).}

\item{color}{String. R Color.}
}
\description{
Visualizes the event density for a given patient's data using the Coxmos model.
}
\details{
The \code{plot_patient.eventDensity} function provides a graphical representation of the event
density for a specific patient's data, based on the Coxmos model. The function computes the density
of events and non-events and plots them, highlighting the predicted value for the given patient's
data. The density is determined using density estimation, and the predicted value is obtained from
the Coxmos model. The function allows customization of the plot aesthetics, such as point size and
color. The resulting plot provides a visual comparison of the patient's predicted event density
against the overall event density distribution, aiding in the interpretation of the patient's risk
profile.
}
\examples{
\dontrun{
plot_patient.eventDensity(patient, time = NULL, model)
}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
