% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_patient.eventHistogram}
\alias{plot_patient.eventHistogram}
\title{patient.eventHistogram}
\usage{
plot_patient.eventHistogram(
  patient,
  time = NULL,
  model,
  type = "lp",
  size = 3,
  color = "red"
)
}
\arguments{
\item{patient}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation. Qualitative variables must be transform into binary variables.}

\item{time}{Numeric. Time point where the AUC will be evaluated (default: NULL).}

\item{model}{HDcox model.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{size}{Numeric. Point size (default: 3).}

\item{color}{String. R Color.}
}
\description{
Generates a histogram plot for the HDcox model selected, and adds a new point for a
test patient specified in the "patient" parameter. The "type" parameter specifies the type of Cox
prediction to use, with "lp", "risk", "expected", or "survival" as options.
}
\examples{
\dontrun{
plot_patient.eventHistogram(patient, time = NULL, model)
}
}
