% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_patient.eventHistogram}
\alias{plot_patient.eventHistogram}
\title{patient.eventHistogram}
\usage{
plot_patient.eventHistogram(
  patient,
  time = NULL,
  model,
  type = "lp",
  size = 3,
  color = "red"
)
}
\arguments{
\item{patient}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation. Qualitative variables must be transform into binary variables.}

\item{time}{Numeric. Time point where the AUC will be evaluated.}

\item{model}{HDcox model.}

\item{type}{Character. Prediction type: "lp", "risk", "expected" or "survival" (default: "lp").}

\item{size}{Numeric. Point size (default: 3).}

\item{color}{String. R Color.}
}
\description{
patient.eventHistogram
}
