% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_proportionalHazard.list}
\alias{plot_proportionalHazard.list}
\title{plot_proportionalHazard.list}
\usage{
plot_proportionalHazard.list(lst_models)
}
\arguments{
\item{lst_models}{List of HDcox models}
}
\description{
plot_proportionalHazard.list
}
