% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_proportionalHazard}
\alias{plot_proportionalHazard}
\title{plot_proportionalHazard}
\usage{
plot_proportionalHazard(model)
}
\arguments{
\item{model}{HDcox model}
}
\description{
plot_proportionalHazard
}
