% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions.R
\name{plot_proportionalHazard}
\alias{plot_proportionalHazard}
\title{plot_proportionalHazard}
\usage{
plot_proportionalHazard(model)
}
\arguments{
\item{model}{Coxmos model.}
}
\value{
A ggplot2 plot.
}
\description{
Perform the proportional hazard assumption for a specific Coxmos model.
It uses the functions "survival::cox.zph" and "survminer::ggcoxzph", but it returns a ggplot2 graph.
}
\examples{
\dontrun{
plot_proportionalHazard(model)
}
}
\references{
\insertRef{survival_package}{Coxmos}
\insertRef{survminer_package}{Coxmos}
\insertRef{Grambsch_1994}{Coxmos}
\insertRef{Schoenfeld_1982}{Coxmos}
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
