% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_proportionalHazard}
\alias{plot_proportionalHazard}
\title{plot_proportionalHazard}
\usage{
plot_proportionalHazard(model)
}
\arguments{
\item{model}{HDcox model.}
}
\value{
A ggplot2 plot.
}
\description{
Performs a test of the proportional hazards assumption for a HDcox model,
and generates a ggplot2 to visualize the results. The function uses the "cox.zph" function
from the "survival" package and the "ggcoxzph" function from the "survminer" package to
generate the plot. The "model" parameter is the HDcox model to be tested. The plot shows
the scaled Schoenfeld residuals for each variable/component included in the model, plotted against
time. If the proportional hazards assumption holds, the residuals should be randomly distributed
around zero.
}
\examples{
\dontrun{
plot_proportionalHazard(model)
}
}
