% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta.list}
\alias{plot_pseudobeta.list}
\title{plot_pseudobeta.list}
\usage{
plot_pseudobeta.list(
  lst_models,
  error.bar = T,
  onlySig = F,
  alpha = 0.05,
  zero.rm = T,
  top = NULL,
  auto.limits = T,
  show_percentage = T,
  size_percentage = 3,
  verbose = F
)
}
\arguments{
\item{lst_models}{List of Coxmos models.}

\item{error.bar}{Logical. Show error bar (default: TRUE).}

\item{onlySig}{Logical. Compute pseudobetas using only significant components (default: FALSE).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0 (default: TRUE).}

\item{top}{Numeric. Show "top" first variables with the higher pseudobetas in absolute value. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{show_percentage}{Logical. If show_percentage = TRUE, it shows the contribution percentage for each variable to the full model (default: TRUE).}

\item{size_percentage}{Numeric. Size of percentage text (default: 3).}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}
}
\description{
Run the function "plot_pseudobeta" for a list of models. More information in "?plot_pseudobeta".
}
\examples{
\dontrun{
plot_pseudobeta.list(lst_models)
}
}
