% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta}
\alias{plot_pseudobeta}
\title{plot_pseudobeta}
\usage{
plot_pseudobeta(
  model,
  error.bar = T,
  onlySig = F,
  alpha = 0.05,
  zero.rm = F,
  top = NULL,
  auto.limits = T,
  show_percentage = T,
  size_percentage = 3
)
}
\arguments{
\item{model}{HDcox model.}

\item{error.bar}{Logical. Show error bar (default: TRUE).}

\item{onlySig}{Logical. Compute psudobetas using only significant components.}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0.}

\item{top}{Numeric. Show "top" first variables with the higher pseudobetas in absolute value. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{show_percentage}{Logical. If show_percentage = TRUE, it shows the contribution percentage for each variable to the full model (default: TRUE).}

\item{size_percentage}{Numeric. Size of percentage text (default: 3).}
}
\description{
plot_pseudobeta
}
