% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta_newPatient}
\alias{plot_pseudobeta_newPatient}
\title{plot_pseudobeta.newPatient}
\usage{
plot_pseudobeta_newPatient(
  model,
  new_observation,
  error.bar = T,
  onlySig = T,
  alpha = 0.05,
  zero.rm = T,
  top = NULL,
  auto.limits = T,
  show.betas = F
)
}
\arguments{
\item{model}{HDcox model.}

\item{new_observation}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation. Qualitative variables must be transform into binary variables.}

\item{error.bar}{Logical. Show error bar (default: TRUE).}

\item{onlySig}{Logical. Compute psudobetas using only significant components (default: TRUE).}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0 (default: TRUE).}

\item{top}{Numeric. Show "top" first variables with the higher pseudobetas in absolute value. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{show.betas}{Logical. Show original betas (default: FALSE).}
}
\description{
This function generates a ggplot2 graph that compares the pseudobeta direction from
the HDcox model with the values of a new observation. The beta coefficients are shown as bars on
the right-hand side of the plot, where a positive coefficient indicates a risky variable and a
negative coefficient indicates a protective variable. On the left-hand side of the plot, another
bar plot is shown with the same variables for the new patient. The direction of each bar is either
up or down, depending on the value of the new patient. If the direction of the bar for a variable
is the same as the pseudobeta coefficient, it indicates that the variable is dangerous for the new
patient. The function also allows for the inclusion of error bars, only significant variables, top
variables, and custom alpha and size values.
}
\examples{
\dontrun{
plot_pseudobeta_newPatient(model, new_observation)
}
}
