% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta_newPatient}
\alias{plot_pseudobeta_newPatient}
\title{plot_pseudobeta.newPatient}
\usage{
plot_pseudobeta_newPatient(
  model,
  new_observation,
  error.bar = T,
  onlySig = T,
  alpha = 0.05,
  zero.rm = T,
  top = NULL,
  auto.limits = T,
  show.betas = F
)
}
\arguments{
\item{model}{HDcox model.}

\item{new_observation}{Numeric matrix or data.frame. New explanatory variables (raw data) for one observation. Qualitative variables must be transform into binary variables.}

\item{error.bar}{Logical. Show error bar (default: TRUE).}

\item{onlySig}{Logical. Compute psudobetas using only significant components.}

\item{alpha}{Numeric. Numerical values are regarded as significant if they fall below the threshold (default: 0.05).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0.}

\item{top}{Numeric. Show "top" first variables with the higher pseudobetas in absolute value. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}

\item{show.betas}{Logical. Show original betas (default: FALSE).}
}
\description{
plot_pseudobeta.newPatient
}
