% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_pseudobeta_newPatient}
\alias{plot_pseudobeta_newPatient}
\title{plot_pseudobeta.newPatient}
\usage{
plot_pseudobeta_newPatient(
  model,
  new_pat,
  error.bar = T,
  onlySig = T,
  alpha = 0.05,
  zero.rm = T,
  top = NULL,
  auto.limits = T,
  show.betas = F
)
}
\arguments{
\item{model}{HDcox model}

\item{new_pat}{Row of new patients}

\item{error.bar}{Logical. Show error bar}

\item{onlySig}{Logical. Compute psudobetas using only significant components.}

\item{alpha}{Significant value (P.value).}

\item{zero.rm}{Logical. Remove variables with a pseudobeta equal to 0.}

\item{top}{Plot the top X variables with the higher pseudobetas in absolute value.}

\item{auto.limits}{Compute Y limit automatically}

\item{show.betas}{Show original betas}
}
\description{
plot_pseudobeta.newPatient
}
