% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_time.list}
\alias{plot_time.list}
\title{Time consuming plot.}
\usage{
plot_time.list(lst_models, x.text = "Method", y.text = NULL)
}
\arguments{
\item{lst_models}{List of HDcox models. Each HDCox object has the attribute time measured in minutes (cross-validation models could be also added to this function).}

\item{x.text}{Character. X axis title.}

\item{y.text}{Character. Y axis title. If y.text = NULL, then y.text = "Time (mins)" (default: NULL).}
}
\value{
A ggplot2 bar plot.
}
\description{
Generates a ggplot2 with the consumed time for each model included in the list of
HDcox models passed. The list could include cox and cross-validation HDcox models (contain the
variable "time" inside the model objects). The plot displays the consumed time on the y-axis and
the names of the models on the x-axis. This method requires the ggplot2 package to be installed.
}
\details{
For time comparison between best models, use the cross validation objects instead the final models. The training has to be taken into account.
}
\examples{
\dontrun{
  lst_models = list("cox" = cox.model, "cv.sPLS-ICOX" = cv.splsicox_model,
  "sPLS-ICOX" = splsicox_model)
  plot_time.list(lst_models, x.text = "Method")
}
}
