% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{plot_time.list}
\alias{plot_time.list}
\title{Time consuming plot.}
\usage{
plot_time.list(lst_models, x.text = "Method", y.text = NULL)
}
\arguments{
\item{lst_models}{List of HDcox models. Each HDCox object has the attribute time measured in minutes.}

\item{x.text}{Character. X axis title.}

\item{y.text}{Character. Y axis title. If y.text = NULL, then y.text = "Time (mins)" (default: NULL).}
}
\value{
A ggplot2 bar plot.
}
\description{
Generate a ggplot2 with the consumed time for each model included in "lst_models". The models have to be generated by the HDcox methods (include the variable time inside the model objects).
}
\details{
For time comparison between best models, use the cross validation objects instead the final models. The training has to be taken into account.
}
\examples{
\dontrun{
  lst_models = {"cox" = cox.model, "cv.sPLS-ICOX" = cv.splsicox_model, "sPLS-ICOX" = splsicox_model}
  plot_time.list(lst_models, x.text = "Method")
}
}
