% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{predict.HDcox}
\alias{predict.HDcox}
\title{predict.HDcox}
\usage{
\method{predict}{HDcox}(object, ..., newdata = NULL)
}
\arguments{
\item{object}{HDcox model}

\item{...}{additional arguments affecting the predictions produced.}

\item{newdata}{Numeric matrix or data.frame. New data for explanatory variables (raw data). Qualitative variables must be transform into binary variables.}
}
\value{
Score values for new data using the HDcox model selected.
}
\description{
predict.HDcox
}
