% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{predict.HDcox}
\alias{predict.HDcox}
\title{predict.HDcox}
\usage{
\method{predict}{HDcox}(object, ..., newdata = NULL)
}
\arguments{
\item{object}{HDcox model}

\item{...}{additional arguments affecting the predictions produced.}

\item{newdata}{New data (X matrix)}
}
\value{
Score values for new data using the PLS model.
}
\description{
predict.HDcox
}
