% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_common_functions.R
\name{print.Coxmos}
\alias{print.Coxmos}
\title{print.Coxmos}
\usage{
\method{print}{Coxmos}(x, ...)
}
\arguments{
\item{x}{Coxmos object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The function is used to print the output of an object of class Coxmos. The function
prints the final cox model if it is a survival model, or the characteristics of the best model if
it is a cross-validated model.
}
\examples{
\dontrun{
cv.model <- cv.splsicox(X,Y, max.ncomp = 8, spv_penalty.list = seq(0.1,0.9,0.2), n_run = 5,
k_folds = 10)
print.Coxmos(cv.model)

model <- splsicox(X,Y)
print.Coxmos(model)
}
}
