% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{save_ggplot.svg}
\alias{save_ggplot.svg}
\title{save_ggplot.svg}
\usage{
save_ggplot.svg(
  plot,
  folder = NULL,
  name = "plot",
  wide = T,
  quality = "4K",
  dpi = 80,
  custom = NULL
)
}
\arguments{
\item{plot}{ggplot2 object}

\item{folder}{folder}

\item{name}{file name}

\item{wide}{16:9 or 4:3}

\item{quality}{one of: "HD", "FHD", "2K", "4K", "8K"}

\item{dpi}{dpi}

\item{custom}{Custom size. Numeric vector of width and height}
}
\description{
Allows to save ggplot2 objects in .svg format based on an specific resolution.
}
\examples{
\dontrun{
save_ggplot.svg(plot, folder)
}
}
