% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_splsdrcox.R
\name{splsdrcox}
\alias{splsdrcox}
\title{sPLS-DRCOX}
\usage{
splsdrcox(
  X,
  Y,
  n.comp = 2,
  eta = 0.5,
  x.center = TRUE,
  x.scale = FALSE,
  y.center = FALSE,
  y.scale = FALSE,
  remove_near_zero_variance = T,
  remove_zero_variance = F,
  toKeep.zv = NULL,
  MIN_EPV = 5,
  returnData = T,
  verbose = F
)
}
\arguments{
\item{X}{Numeric matrix. Predictor variables}

\item{Y}{Numeric matrix. Response variables. It assumes it has two columns named as "time" and "event". For event column, values can be 0/1 or FALSE/TRUE for censored and event samples.}

\item{n.comp}{Numeric. Number of principal components to compute in the PLS model.}

\item{eta}{Numeric [0-1). Penalty for sPLS. Mean 0 no penalty and 1 maximum penalty. Greater than 1 cannot be selected.}

\item{x.center}{Logical. If x.center = TRUE, X matrix is centered to zero means (default: TRUE).}

\item{x.scale}{Logical. If x.scale = TRUE, X matrix is scaled to unit variances (default: FALSE).}

\item{y.center}{Logical. If y.center = TRUE, Y matrix is centered to zero means (default: FALSE).}

\item{y.scale}{Logical. If y.scale = TRUE, Y matrix is scaled to unit variances (default: FALSE).}

\item{remove_near_zero_variance}{Logical. If remove_near_zero_variance = TRUE, remove_near_zero_variance variables will be removed.}

\item{remove_zero_variance}{Logical. If remove_zero_variance = TRUE, remove_zero_variance variables will be removed.}

\item{toKeep.zv}{Character vector. Name of variables in X to not be deleted by (near) zero variance filtering.}

\item{MIN_EPV}{Minimum number of Events Per Variable you want reach for the final cox model. Used to restrict the number of variables can appear in cox model. If the minimum is not meet, the model is not computed.}

\item{returnData}{Logical. Return original and normalized X and Y matrices.}

\item{verbose}{Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).}
}
\value{
Instance of class "HDcox" and model "sPLS-DRCOX". The class contains the following elements:
\code{X}: List of normalized X data information.
\itemize{
\item \code{(data)}: normalized X matrix
\item \code{(weightings)}: PLS weights
\item \code{(weightings_norm)}: PLS normalize weights
\item \code{(W.star)}: PLS W* vector
\item \code{(scores)}: PLS scores/variates
\item \code{(E)}: error matrices
\item \code{(x.mean)}: mean values for X matrix
\item \code{(x.sd)}: standard deviation for X matrix
}
\code{Y}: List of normalized Y data information.
\itemize{
\item \code{(deviance_residuals)}: deviance residual vector used as Y matrix in the sPLS.
\item \code{(dr.mean)}: mean values for deviance residuals Y matrix
\item \code{(dr.sd)}: standard deviation for deviance residuals Y matrix'
\item \code{(data)}: normalized X matrix
\item \code{(y.mean)}: mean values for Y matrix
\item \code{(y.sd)}: standard deviation for Y matrix'
}
\code{survival_model}: List of survival model information.
\itemize{
\item \code{fit}: coxph object.
\item \code{AIC}: AIC of cox model.
\item \code{BIC}: BIC of cox model.
\item \code{lp}: linear predictors for train data.
\item \code{coef}: Coefficients for cox model.
\item \code{YChapeau}: Y Chapeau residuals.
\item \code{Yresidus}: Y residuals.
}

\code{eta}: Penalty value selected.

\code{n.comp}: Number of components selected.

\code{var_by_component}: Variables selected in each PLS component.

\code{call}: call function

\code{X_input}: X input matrix

\code{Y_input}: Y input matrix

\code{B.hat}: PLS beta matrix

\code{R2}: PLS R2

\code{SCR}: PLS SCR

\code{SCT}: PLS SCT

\code{nzv}: Variables removed by remove_near_zero_variance or remove_zero_variance.

\code{time}: time consumed for running the cox analysis.
}
\description{
Performs a sPLS-DRCOX model (based on plsRcox R package idea).
}
