% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_comparison_functions.R
\name{super.trainAllModels}
\alias{super.trainAllModels}
\title{super.trainAllModels}
\usage{
super.trainAllModels(
  lst_subdata,
  methods,
  comp_calculation = "manual",
  ncomp = 5,
  EN.alpha = 0.5,
  eta = 0.5,
  x.center = TRUE,
  x.scale = TRUE,
  MIN_EPV = 0,
  PARALLEL = TRUE
)
}
\arguments{
\item{lst_subdata}{List of data (X, Y)}

\item{methods}{Methods to use for train}

\item{comp_calculation}{Compute optimal number of components by "manual" or "auto".}

\item{ncomp}{N.components for manual detection.}

\item{EN.alpha}{Penalization for manual detection,}

\item{eta}{Eta for manual detection,}

\item{x.center}{Logical. If x.center = TRUE, X matrix is centered to zero means (default: TRUE).}

\item{x.scale}{Logical. If x.scale = TRUE, X matrix is scaled to unit variances (default: FALSE).}

\item{MIN_EPV}{Numeric. Minimum number of Events Per Variable (EPV) you want reach for the final
cox model. Used to restrict the number of variables/components can be computed in final cox models.
If the minimum is not meet, the model cannot be computed (default: 5).}

\item{PARALLEL}{Logical. Run the cross validation with multicore option. As many cores as your
total cores - 1 will be used. It could lead to higher RAM consumption (default: FALSE).}
}
\description{
super.trainAllModels
}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
