% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxmos_plot_functions_others.R
\name{w.starplot.Coxmos}
\alias{w.starplot.Coxmos}
\title{w.starplot.Coxmos}
\usage{
w.starplot.Coxmos(model, zero.rm = F, top = NULL, auto.limits = T)
}
\arguments{
\item{model}{Coxmos model.}

\item{zero.rm}{Logical. Remove variables equal to 0 (default: FALSE).}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}
}
\description{
w.starplot.Coxmos
}
