% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDcox_plot_functions.R
\name{w.starplot.HDcox}
\alias{w.starplot.HDcox}
\title{w.starplot.HDcox}
\usage{
w.starplot.HDcox(model, zero.rm = F, top = NULL, auto.limits = T)
}
\arguments{
\item{model}{HDcox model.}

\item{zero.rm}{Logical. Remove variables equal to 0.}

\item{top}{Numeric. Show "top" first variables. If top = NULL, all variables are shown (default: NULL).}

\item{auto.limits}{Logical. If "auto.limits" = TRUE, limits are detected automatically (default: TRUE).}
}
\description{
w.starplot.HDcox
}
